+++ aclogview/Tools/OptionsForm.cs
@@ -1,4 +1,4 @@
-﻿using aclogview.Properties;
+using aclogview.Properties;
 using System;
 using System.Security.Policy;
 using System.Windows.Forms;
@@ -11,6 +11,7 @@ namespace aclogview
         private byte _protocolUpdateIntervalDays;
         private bool _parsedDataTreeviewDisplayTooltips;
         private byte _packetsListviewTimeFormat;
+        private byte _treeviewCopyAllPadding;
 
         public OptionsForm()
         {
@@ -20,12 +21,14 @@ namespace aclogview
         private void Options_Load(object sender, EventArgs e)
         {
             LoadDefaultSettings();
+            // Now update the form with our settings
             textBox_ProtocolDays.Text = _protocolUpdateIntervalDays.ToString();
             checkBox_ProtocolUpdates.Checked = _protocolCheckForUpdates;
             if (!_protocolCheckForUpdates)
                 textBox_ProtocolDays.Enabled = false;
             displayNodeTooltips.Checked = _parsedDataTreeviewDisplayTooltips;
             comboBox_TimeFormat.SelectedIndex = _packetsListviewTimeFormat;
+            copyAllPadding.Text = _treeviewCopyAllPadding.ToString();
         }
 
         private void LoadDefaultSettings()
@@ -35,6 +38,7 @@ namespace aclogview
             _protocolCheckForUpdates = Settings.Default.ProtocolCheckForUpdates;
             _parsedDataTreeviewDisplayTooltips = Settings.Default.ParsedDataTreeviewDisplayTooltips;
             _packetsListviewTimeFormat = Settings.Default.PacketsListviewTimeFormat;
+            _treeviewCopyAllPadding = Settings.Default.TreeviewCopyAllPadding;
         }
 
         private void button_OK_Click(object sender, EventArgs e)
@@ -49,7 +53,8 @@ namespace aclogview
         {
             var result = FillProtocolSettings();
             if (!result) return false;
-            FillTreeviewSettings();
+            result = FillTreeviewSettings();
+            if (!result) return false;
 
             return true;
         }
@@ -60,12 +65,30 @@ namespace aclogview
             Settings.Default.ProtocolCheckForUpdates = _protocolCheckForUpdates;
             Settings.Default.ParsedDataTreeviewDisplayTooltips = _parsedDataTreeviewDisplayTooltips;
             Settings.Default.PacketsListviewTimeFormat = _packetsListviewTimeFormat;
+            Settings.Default.TreeviewCopyAllPadding = _treeviewCopyAllPadding;
             Settings.Default.Save();
         }
 
-        private void FillTreeviewSettings()
+        private bool FillTreeviewSettings()
         {
             _parsedDataTreeviewDisplayTooltips = displayNodeTooltips.Checked;
+            var isValidInput = ValidateCopyAllPadding();
+            return isValidInput && ParseCopyAllPadding();
+        }
+
+        private bool ValidateCopyAllPadding()
+        {
+            if (!string.IsNullOrWhiteSpace(copyAllPadding.Text)) return true;
+            DisplayInvalidPaddingTooltip();
+            return false;
+        }
+
+        private bool ParseCopyAllPadding()
+        {
+            var result = byte.TryParse(copyAllPadding.Text, out _treeviewCopyAllPadding);
+            if (result) return true;
+            DisplayByteConversionError("Pad \"Copy All\" output field");
+            return false;
         }
 
         private bool FillProtocolSettings()
@@ -87,11 +110,11 @@ namespace aclogview
         {
             var result = byte.TryParse(textBox_ProtocolDays.Text, out _protocolUpdateIntervalDays);
             if (result) return true;
-            DisplayByteConversionError(textBox_ProtocolDays.Text);
+            DisplayByteConversionError("protocol documentation update interval");
             return false;
         }
 
-        private void DisplayInvalidDaysTooltip ()
+        private void DisplayInvalidDaysTooltip()
         {
             toolTip_ProtocolDoc.ToolTipTitle = "Invalid Input";
             toolTip_ProtocolDoc.Show("You must enter a valid number of days for the update interval.",
@@ -100,10 +123,18 @@ namespace aclogview
                 5000);
         }
 
-        private static void DisplayByteConversionError (string updateDays)
+        private void DisplayInvalidPaddingTooltip()
         {
-            MessageBox.Show("Could not convert the protocol documentation " +
-                        $"update interval string '{updateDays}' to a byte.",
+            toolTip_ProtocolDoc.ToolTipTitle = "Invalid Input";
+            toolTip_ProtocolDoc.Show("You must enter a valid number padding spaces.",
+                this,
+                copyAllPadding.Location,
+                5000);
+        }
+
+        private static void DisplayByteConversionError(string field)
+        {
+            MessageBox.Show($"The {field} must be a number between 0-255.",
                         "Error:",
                         MessageBoxButtons.OK,
                         MessageBoxIcon.Error);